modify deleteImageSubmit and add isImageOwner Method.

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -150,13 +150,34 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, HttpSession session, RedirectAttributes redirectAttrs, Model model) {
+
+        if( isImageOwner(imageId, session) ) {
+            imageService.deleteImage(imageId);
+            return "redirect:/images";
+        } else {
+            String imageTitle = imageService.getImage(imageId).getTitle();
+            String error = "Only the owner of the image can delete the image";
+            redirectAttrs.addAttribute("deleteError", error).addFlashAttribute("deleteError", error);
+            return "redirect:/images/" + imageId + '/' + imageTitle;
+        }
+
+
     }
 
+    private Boolean isImageOwner(Integer imageId, HttpSession session){
+            Image currentImageData = imageService.getImage(imageId);
+            Integer currentImageOwnerId = currentImageData.getUser().getId();
 
-    //This method converts the image to Base64 format
+            User loggedInUser = (User) session.getAttribute("loggeduser");
+            Integer loggedInUserId = loggedInUser.getId();
+
+            return loggedInUserId.equals(currentImageOwnerId);
+        }
+
+
+
+        //This method converts the image to Base64 format
     private String convertUploadedFileToBase64(MultipartFile file) throws IOException {
         return Base64.getEncoder().encodeToString(file.getBytes());
     }
